requirements.txt
Babel==2.6.0
Flask-BabelEx==0.9.3
Flask-Caching==1.7.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Testing==0.7.1
Flask-WTF==0.14.2
Flask==1.0.2
SQLAlchemy==1.3.3
WTForms==2.2.1
Werkzeug==0.15.2
apache-libcloud==2.4.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
cached-property==1.5.1
celery[sqlalchemy]==4.3.0
environs==4.1.2
gunicorn==19.9.0
mkwvconf==0.1.1
passlib==1.7.1
python-crontab==2.3.6
requests==2.21.0
typing==3.6.6
tzlocal==1.5.1
xtarfile[zstd]==0.0.2
